\name{GADM36SF}
\alias{GADM36SF}
\docType{data}
\title{data.frame of maps provided by gadm_org)}
\description{
Dataset of description of all maps provided by gadm_org. This has been used to generate the vignette GADMTools_ISO_3166-1_alpha-3}
\usage{data(GADM36SF)}
\format{
  A data.frame.
}
\keyword{datasets}
