\name{GADMTools-package}
\alias{GADMTools-package}
\alias{GADMTools}
\docType{package}
\title{
Easy use of GADM shapefiles
}
\description{
See; \url{https://gadm.org/}

\bold{GADM} is a spatial database of the world's adminstrative boundaries for use in \bold{GIS} and similar software. Administrative areas in this database are countries and lower level subdivisions such as provinces, departments, cantons, etc.

With \bold{GADMTools}, a wrapper for \bold{GADM} shapefiles, you can easily manipulate, assemble, and create subsets of these objects.
}
\details{
\tabular{ll}{
Package: \tab GADMTools\cr
Type: \tab Package\cr
Version: \tab 2.2-1\cr
Date: \tab 2018-06-04\cr
License: \tab GPL-3\cr
}

}
\author{
Jean Pierre Decorps <jp.decorps@epiconcept.fr>

Maintainer: 
Jean Pierre Decorps <jp.decorps@epiconcept.fr>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
%%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
## NOTHING
}
