\name{grid.map}
\alias{grid.map}

\title{ Arrange maps on a grid }

\description{
.
}
\usage{
grid.map(left, right, center=NULL, title=NULL)
}
\arguments{
  \item{left}{ \bold{Object} GADMWrapper }
  
  \item{right}{
\bold{data.frame} - data to plot
}
  
  \item{center}{
  \bold{String} - A RColorBrewer palette name or a \bold{String vector} a
  vector of colors. Default \bold{NULL}.
}
  \item{title}{
  \bold{String} - Title of the plot. Default is an empty string
}
}
\details{
----
}
\value{
    \bold{Object}  ggplot2
}
\references{
----
}
\author{
Jean Pierre Decorps <jp.decorps@epiconcept.fr>
}
\note{
----
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
----
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--  or do  help(data=index)  for the standard data sets.

}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
