\name{GADMTools-package}
\alias{GADMTools-package}
\alias{GADMTools}
\docType{package}
\title{
Easy use of GADM shapefiles
}
\description{
See; \url{https://gadm.org/}

\bold{GADM} is a spatial database of the world's adminstrative boundaries for use in \bold{GIS} and similar software. Administrative areas in this database are countries and lower level subdivisions such as provinces, departments, cantons, etc.

With \bold{GADMTools}, a wrapper for \bold{GADM} shapefiles, you can easily manipulate, assemble, and create subsets of these objects.

\bold{GADMTools} can use 2 shapefile formats,\bold{SpatialPolyonsDataFrame} and \bold{Simple Features (SF)}, both provided by GADM as .rds files.

\bold{NB:} the SF format is supported only from version 3.5 of GADMTools.

}
\details{
\tabular{ll}{
Package: \tab GADMTools\cr
Type: \tab Package\cr
Version: \tab 3.7-1\cr
Date: \tab 2019-06-25\cr
License: \tab GPL-3\cr
}

}
\author{
Jean Pierre Decorps <jp.decorps@epiconcept.fr>

Maintainer: 
Jean Pierre Decorps <jp.decorps@epiconcept.fr>
}
